refactor: group publisher packages under internal/publishers/ [sctl-7mk]
