+++ internal/publishers/factory_test.go
@@ -4,12 +4,12 @@ import (
 	"testing"
 
 	"github.com/lumberbarons/solar-controller/internal/config"
-	"github.com/lumberbarons/solar-controller/internal/file"
-	"github.com/lumberbarons/solar-controller/internal/mqtt"
 	"github.com/lumberbarons/solar-controller/internal/publish"
-	"github.com/lumberbarons/solar-controller/internal/remotewrite"
-	"github.com/lumberbarons/solar-controller/internal/sns"
-	"github.com/lumberbarons/solar-controller/internal/solace"
+	"github.com/lumberbarons/solar-controller/internal/publishers/file"
+	"github.com/lumberbarons/solar-controller/internal/publishers/mqtt"
+	"github.com/lumberbarons/solar-controller/internal/publishers/remotewrite"
+	"github.com/lumberbarons/solar-controller/internal/publishers/sns"
+	"github.com/lumberbarons/solar-controller/internal/publishers/solace"
 )
 
 func TestNewPublisher(t *testing.T) {
