+++ internal/publishers/file/publisher.go
