+++ internal/publishers/solace/publisher.go
