+++ internal/publishers/solace/publisher_test.go
